Move createPncAdjudication function into root lib folder
